Use Passer interface for MQTT over WS
