+++ go.mod
@@ -5,13 +5,13 @@ go 1.21
 toolchain go1.21.4
 
 require (
-	github.com/caarlos0/env/v11 v11.0.0
-	github.com/eclipse/paho.mqtt.golang v1.4.3
+	github.com/caarlos0/env/v11 v11.2.0
+	github.com/eclipse/paho.mqtt.golang v1.5.0
 	github.com/google/uuid v1.6.0
-	github.com/gorilla/websocket v1.5.1
+	github.com/gorilla/websocket v1.5.3
 	github.com/joho/godotenv v1.5.1
-	golang.org/x/crypto v0.22.0
-	golang.org/x/sync v0.7.0
+	golang.org/x/crypto v0.25.0
+	golang.org/x/sync v0.8.0
 )
 
-require golang.org/x/net v0.24.0 // indirect
+require golang.org/x/net v0.27.0 // indirect
